requirement.txt
asgiref==3.5.0
certifi==2023.7.22
distlib==0.3.7
Django==4.0.3
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.13.1
packaging==23.2
pipenv==2023.10.24
platformdirs==3.11.0
pytesseract==0.3.10
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.24.6
